3-puts

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -8,7 +8,7 @@
 */
 void _puts(char *str)
 {
-		long unsigned int tmp = 0;
+		unsigned int tmp = 0;
 
 		while (tmp < strlen(str))
 		{
